+++ r2d2/src/lib.rs
@@ -22,7 +22,9 @@
 
 mod manager;
 
-pub use deadpool::managed::sync::reexports::*;
-pub use deadpool::managed::{sync::SyncWrapper, Pool, PoolError};
+pub use deadpool::managed::reexports::*;
+pub use deadpool::managed::{Pool, PoolError};
+pub use deadpool_sync::reexports::*;
+pub use deadpool_sync::SyncWrapper;
 
 pub use self::manager::Manager;