+++ diesel/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## v0.3.0 (unreleased)
+
+* __Breaking:__ Replace `deadpool::managed::sync` by
+  `deadpool-sync::SyncWrapper` which fixes the return type
+  of the `interact` method.
+
 ## v0.2.0
 
 * __Breaking:__ Replace `config` feature with `serde` (opted out by default)