+++ r2d2/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## v0.2.0 (unreleased)
+
+* __Breaking:__ Replace `deadpool::managed::sync` by
+  `deadpool-sync::SyncWrapper` which fixes the return type
+  of the `interact` method.
+
 ## v0.1.0
 
 * First release